Rework data downloading

+++ SIT.Manager.Avalonia/Services/FileService.cs
@@ -96,7 +96,7 @@ namespace SIT.Manager.Avalonia.Services
                     File.Delete(filePath);
                 }
 
-                var progress = new Progress<float>((prog) => {
+                var progress = new Progress<double>((prog) => {
                     _actionNotificationService.UpdateActionNotification(new ActionNotification($"Downloading '{fileName}'", Math.Floor(prog), showProgress));
                 });
